producto personalizabel double

+++ src/main/java/es/upm/iwsim22_01/models/PersonalizableProduct.java
@@ -3,7 +3,7 @@ package es.upm.iwsim22_01.models;
 public class PersonalizableProduct extends Product{
     int maxPers;
 
-    public PersonalizableProduct(int id, String name, Category category, int price, int maxPers){
+    public PersonalizableProduct(int id, String name, Category category, double price, int maxPers){
         super(id, name, category, price);
         this.maxPers = maxPers;
     }
